+++ libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: mviinika <mviinika@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/08 14:13:57 by mviinika          #+#    #+#             */
-/*   Updated: 2022/03/25 16:18:00 by mviinika         ###   ########.fr       */
+/*   Updated: 2022/04/13 09:16:57 by mviinika         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -89,7 +89,7 @@ void	ft_lstrev(t_list **alst);
 int		ft_isspace(int c);
 void	ft_swap(int *a, int *b);
 char	*ft_strndup(const char *s, size_t size);
-int		ft_abs(int num);
+size_t	ft_abs(int num);
 double	ft_pow(double num, double pow);
 int		bintodec(signed long long bin);
 size_t	ft_intlen(long long n);
